A game developed in Python with Kivy framework
This game follows the story of Litlle Redhood, she is on a quest to save her Grandma from the Wolf and the allies: the Evil Crow and the Vicious Rattlesnake.
To pass each level, the user (Little Redhood pov) shoots different projectiles to many various obstacles. At each level there is a different main enemy, whose defeat means passing the level. 
To win the game, one has to overcome the Wolf, hence Grandma is saved.